requirements.txt
Flask==2.3.3
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0
gunicorn==21.2.0